Split into own file and address concerns
