+++ main.go
@@ -204,37 +204,3 @@ func doError(rw http.ResponseWriter, req *http.Request, err error) {
 	log.Printf("ERROR: %s", err.Error())
 	rw.WriteHeader(500)
 }
-
-func Logger(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(rw http.ResponseWriter, req *http.Request) {
-		begin := time.Now()
-		originalPath := req.URL.Path
-		recorder := &resRecorder{ResponseWriter: rw}
-		next.ServeHTTP(recorder, req)
-		log.Printf("%s %s => %d. In %f seconds. Path Rewrite to %s. Cache: %s",
-			req.Method,
-			originalPath,
-			recorder.status,
-			time.Since(begin).Seconds(),
-			req.URL.Path,
-			recorder.header.Get("X-Cache"),
-		)
-	})
-}
-
-type resRecorder struct {
-	http.ResponseWriter
-	status int
-	header http.Header
-}
-
-func (rr *resRecorder) Header() http.Header {
-	hdr := rr.ResponseWriter.Header()
-	rr.header = hdr
-	return hdr
-}
-
-func (rr *resRecorder) WriteHeader(status int) {
-	rr.status = status
-	rr.ResponseWriter.WriteHeader(status)
-}
